Revert "Modified Direction, Position, and Moving Game Element to use floats again instead of integers."
This reverts commit 6e459025

+++ app/src/main/java/com/twilightcitizen/whackapede/gameElements/Direction.java
@@ -9,8 +9,8 @@ package com.twilightcitizen.whackapede.gameElements;
 
 public class Direction {
     // Components along the X and Y axes.
-    private final float x;
-    private final float y;
+    private final int x;
+    private final int y;
 
     // Static Exit instances for exiting a turn from the Top, Bottom, Left, or Right.
     public static final Direction up    = new Direction(  0, -1 );
@@ -18,9 +18,9 @@ public class Direction {
     public static final Direction left  = new Direction( -1,  0 );
     public static final Direction right = new Direction(  1,  0 );
 
-    public Direction( float x, float y ) { this.x = x; this.y = y; }
+    public Direction( int x, int y ) { this.x = x; this.y = y; }
 
     // Components along the X and Y axes are read-only.
-    public float getX() { return x; }
-    public float getY() { return y; }
+    public int getX() { return x; }
+    public int getY() { return y; }
 }
