+++ app/src/main/java/com/twilightcitizen/whackapede/gameElements/MovingGameElement.java
@@ -9,12 +9,12 @@ package com.twilightcitizen.whackapede.gameElements;
 
 public class MovingGameElement extends GameElement {
     // Speed of traversal.
-    private float speed;
+    private int speed;
 
     // Direction along the X and Y axes.
     private Direction direction;
 
-    protected MovingGameElement( Position position, float speed, Direction direction ) {
+    protected MovingGameElement( Position position, int speed, Direction direction ) {
         super( position );
 
         this.speed = speed;
@@ -22,8 +22,8 @@ public class MovingGameElement extends GameElement {
     }
 
     // Speed of traversal is read/write.
-    public float getSpeed() { return speed; }
-    public void setSpeed( float speed ) { this.speed = speed; }
+    public int getSpeed() { return speed; }
+    public void setSpeed( int speed ) { this.speed = speed; }
 
     // Direction along the X and Y axes is read/write.
     public Direction getDirection() { return direction; }
